internal/precompiles/ai/python_files/requirements.txt
accelerate==0.31.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
diffusers==0.29.1
filelock==3.15.4
flash-attention==1.0.0
fsspec==2024.6.0
huggingface-hub==0.23.4
idna==3.7
importlib-metadata==7.2.0
jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==24.1
peft==0.11.1
pillow==10.3.0
psutil==6.0.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
sacremoses==0.1.1
safetensors==0.4.3
sentencepiece==0.2.0
sympy==1.12.1
timm==1.0.7
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
typing-extensions==4.12.2
urllib3==2.2.2
websockets==12.0
zipp==3.19.2
